Change logging propagation to be precalculated
Seems I had more options to propagate (--nodblog, --syslog, --quiet), so I
decided to refactor it to be a pre-calculated QString that gets used all over
rather than repeating myself over and over.  I also added a check for --quiet
to avoid putting it in the commandline twice where we want to be sure it's
there.  (Double quiet totally disables logging).

I also added in the HandleSetVerbose to recalculate the propagated argument
string.