Work-around to handle display of high-numbered cards.
If a card's number was greater than 9, only the first digit was
displayed in some situations and caused confusion with some users.  In
these cases, now display "#" instead.  This is a work-around to avoid
an ABI change in fixes/0.24 as was done in the complete fix in
SHA:95b4d928 in master.