Make EITScanner thread only be started when needed
Currently, an EITScanner thread is started for every source that is capable of
doing EIT scans (DVB/HDHomerun) even if EIT is disabled for that source.  This
means we can have several pointless threads that will do us no good.  Rather
than this, let's only start the thread on capable sources that have EIT
enabled.

Also tweaked the conditional use of the scanner to make sure the scanner
exists to preclude possible NULL pointer crashes.  Moved the check for scanner
first to optimize the execution slightly in the non-EIT case.