Renamed LogPrint() to LOG()
As this name was intended to be a temporary name anyways for use in the
conversion, and since some others have expressed a preference for using LOG()
rather than LogPrint(), I'm changing it now to save trouble later.

For clarity: VERBOSE() and LOG() can coexist.  I didn't want to make one
huge commit with the addition of an argument to VERBOSE(), so I needed to make
a temporary name for during the cutover to allow for the cutover to be
done in smaller chunks.