Fix a regresssion in 45cfaa7a8 related to local image loading.
45cfaa7a8 made itso that all images not starting with a '/'
were loaded in the background without first checking the last
modified time in the foreground first.  This somehow caused an
issue with local relative filenames.  This bugfix changes the
logic a little to check for remote filenames instead of checking
for a '/'.

There still may be a race condition or locking issue which
actually caused the problem seen after 45cfaa7a8, but this
patch fixes the code so that the intent of 45cfaa7a8 is accomplished.