AvFormatDecoder: Reset the video codec on stream changes.
I've been staring blankly at this trying to figure out why it hasn't
caused problems before - only to realise that live tv changes always
force a full teardown of the decoder and the codec never changes for dvd
playback.

Should, and appears to, fix a number of blu ray playback issues when the
bd library switches to a title with a different codec.