Metadata: Parse XBMC-style NFO files.
XBMC has a XML NFO format that goes side-by-side files in media directories.  These XML files contain most of the basic metadata.  This commit adds basic support for parsing the metadata out of them.

I didn't implement everything in the format, yet, and I haven't turned it on for Television stuff (since the format is somewhat ugly in this regard-- there's one file for the series, and one for the episode.  I want to think a bit more about it.).  The format also doesn't implement all of the metadata *we* keep and pull in, but it does have the basics that people care about-- title, subtitle, year, season, episode, plot, inetref, etc.

I'll flesh out the parser a little more to cover cast, director, and other people, and maybe watched/unwatched status too.  Once I'm done with that I'll circle back and get it working for TV shows as well.  Most of the needed stuff is already in the parser, just need to combine the two sets of metadata in a way that makes sense.

Format documentation: http://wiki.xbmc.org/?title=Import_-_Export_Library
Hilarious logging censored by sphery.