Metadata: Support import of Myth-format metadata from an external file.
Superior metadata re-import from an external file.  Uses Myth's native metadata XML format.  Looks for a file named the same as the video file, with a .mxml extension.

This supports *all* our metadata natively, will reconnect any downloaded artwork found in the fanart/banner/coverart/etc. directories, and saves us having to do any re-lookup with the metadata source.  If the images go missing but the video files and .mxml files are still present, the images will be redownloaded but that will be the only traffice with the metadata source.

This also allows ambiguous titles to have a perfect match when being removed from and readded to the library.

I'll add support for writing out the .mxml files upon metadata grab soon.