Add CPU utilisation monitoring (linux only).
The current version is piggybacked onto the jitterometer code (and hence
is playback only - but could easily be moved into one of the 'base'
libraries for more generic use) and extends the jitterometer output as
well as the OSD debug screen.

Utilisation is provided for each core. While overall load is nice, it
doesn't give the real flavour of what's happening when playback is CPU
bound (e.g. the decoder is pegging one core). The output is going to
start looking messy for all of the boys and girls with quad/many core
CPUs and hyperthreading and that will probably need to be tweaked (maybe
overall utilisation plus highlight cores that have utilisation > 90%).

Windows code should be easy enough to produce - not so sure about OS X.