Moved the free of the malloced memory to a common code path
from cppcheck:
[libs/libmythbase/system-unix.cpp:813]: (error) Memory leak: command

This memory still may appear leaked in the child code path (after the fork)
but the execv() call replaces everything in the memory footprint of the
process anyways.  The system will take care of that.

If this doesn't silence cppcheck, we'll need to figure out how to make it
happy, but this code shouldn't actually be leaking anything at this point.