Metadata: Add a MetadataFactory Class.
This is an extremely simple-to-use class that pulls together all the other metadata work I've done over the past year or two.  It's a simple QObject that allows you to input a programInfo *or* videometadata, along with switches to automatically match names and download images.  Then all you do is sit back and wait for events to return with 0, 1, or multiple results.  All the receiver has to do is listen and do what it wants with the returned results.  With this code I was able to add metadata lookup to the PlaybackBox with less than ten additional lines.  Obviously a final implementation will need cleaning up, but user selectable artwork for recordings is now in sight.