libmythmetadata: Add a universal metadata results dialog.
To avoid code duplication in all the places that I'm envisioning allowing metadata lookup (setup of recording rules, video dialog, game dialog, etc), this is a universal metadata results dialog that allows one to display multiple potential matches, including their plot, year, artwork, etc.

There are already versions of this dialog in ex-MythVideo and MythGame, and it was only going to get worse, so add this dialog.

I'll retrofit the Video screens and mythgame to use this dialog in the near future.  When I do that, I'll add default themed versions.