libmythmetadata: Allow the MetadataFactory to lookup a MetadataLookup.
This is needed when an initial lookup returns multiple potential lookups and the user is forced to act on them-- we need to be able to pass the resulting lookup back in.

Bumps API again.  I'd love to be able to tell you this is the last time I will do this today, but it's probably not.