Metadata Factory: handle undetermined recording differently.
The method used to match videos that aren't clearly movies or television doesn't work as accurately for recordings.  Add a new function to handle this (which is used automatically when passed a recinfo or pginfo without a subtitle).

Basically, assume that anything without a subtitle is a generic TV episode, and attempt to look it up at TTVDB.  If that has results, great, it's a generic TV episode.  Set the season and episode to 1 and set the inetref appropriately so that it can be used to match season and episode for future records that *do* have a subtitle.

With (uncommitted) code run against all of my recording rules, this is selecting the correct inetref for every recording rule I have.  There *is* a possible false positive issue when a TV show has the same name as a movie (and you're recording the Movie instead of the TV show) but the code I'm working on allows for one to run a query manually as well, which will allow the user to pick up movie metadata and artwork when set correctly.