Add Support for Season, Episode, and Inetref values to recordings.
This commit constitutes the "scary part," namely the deep dark changes to the many database queries dealing with program data, the serialization of that program data on the protocol, the scheduler, and the various types of bindings.

More succinctly put, this commit does nothing but set the stage for the UI and support application portions that are coming.  Once those parts are committed, you will be able to perform a metadata query on a recording rule as you set it up, set its inetref, and forevermore have it perform supplemental metadata lookups to add this data and more to your recordings.  Completed and properly configured recordings will have season and episode information, as well as statically associated artwork.  The fanart, banner, and coverart for a recording will be browsed for and set according to your preference, rather than being "best effort" or "luck of the draw."  You will be able to use these values in user jobs to name transcoded files in a more conventional way.  Theoretically the PlaybackBox could be expanded to show a season/episode hierarchy.  The possibilities go on and on.

In several days of testing, this appears to be stable and operating as expected, including some uncommitted UI portions.  This *is* a hugely invasive change that touches the order of ProgramInfo data via the protocol (renumbering it), and lots of other things.  Mythweb and all the bindings appear to be working.  Scheduling and both live and recordings are working.

Technically speaking the UI portions of this allowing one to export season and episode numbers and display them wherever programinfo is used are present and working-- they're just not populated with anything of value yet.