Add a MythMetdataLookup program.
This program should be thought of as a companion job to records like mythcommflag or mythtranscode.  It takes programs as an input, performs a lookup online, and then sets various pieces of metadata on a successful return.

For now, it can be run in one of two ways:

./mythmetadatalookup

Will cycle through your entire list of recordings, look them up online, and if matches are found it will set:

* For Movies: Inetref
* For Television: Inetref, Season, and Episode

This mode will not re-run a lookup on a movie that already has an inetref, or a television show that already has inetref, season, and episode.

It can also be run as:

./mythmetadatalookup --chanid 12345 --starttime 07-02-2011T19:30:00

(using any supported date-time format)

Which will unequivocally look up a valid program from that chanid and starttime.

Though this program will work and set correct values in the recorded table, to really show its worth it relies on:

* Themes to be updated to show the season and episode values when relevant
* Recording rule editor changes to allow inetref to be set on recording rules, so that they propagate to the recordings (so that you can set the inetref for "Castle" to TVDB's inetref for "Castle (2009)" and have all your lookups work thereafter)

I will also possibly add a mode later that is a little more "promiscuous" and overrides existing program info values-- say you get a generic episode that you really needed, you could theoretically change the subtitle in the UI and run a lookup on it to get correct, full metadata.  I need to think about how best to accomplish this.