Metadata: Add a Universal Image Search Results dialog.
This dialog is used to display the results of an image search for a particular item, allowing the user to select the one they like.

This is basically the old "artworksel" window that few themes have implemented yet, renamed and made usable everywhere.  It will be used in the edit recording rule dialogs to select artwork the user prefers for a recording rule.