Merged addVerbose() into addLogging() for command-line
Since the functionality is now combined anyways, there will be not likely ever
be a reason to set --verbose without wanting access to the rest of the logging
arguments too.  The one outlier was mythmessage, but it's using --verbose as
a boolean switch anyways, so I added it specifically there.