Metadata Lookup: Look up extra metadata on recording start.
This adds the metadata lookup to the jobqueue in the same way that commflag and transcode are there.  By default, when a recording begins the first thing the jobqueue will do is look up the extra metadata.

As with the other mentioned jobs, metadata lookup for your recordings can be turned off on a per-rule or per-backend basis. (per-rule UI changes to come, remotefile needs to get fixed for me to be able to commit those).