Metadata Lookup: Initial commit of Schedule Editor UI.
This is the Inetref, Season and Episode portion of the Schedule editor UI.  This allows one to assign a inetref to a recording rule as I have documented here:

http://www.mythtv.org/wiki/Enhancing_Recordings_with_Metadata_Lookup

I have commented out the Artwork buttons for now-- until #9895 and #9896 are fixed it's unfinishable.

So, in short, one should now be able to go and perform automatic lookups on their recording rules, and the bits committed over the last few days should now automatically run on recording shortly after they begin and fill in season, episode, etc.

I have updated Terra, MythCenter, and MythCenter-wide for the basics, as well as default and default-wide.  There's a lot more that could be done in those themes.  Current Arclight isn't yet in the theme downloader, but you can get it here:

http://www.fecitfacta.com/Arclight.tar.gz

Arclight is also (currently) the only theme that actually displays the season and episode in the PBB (it's in the subtitle field on the left).  You can test that you have everything working by running "mythmetadatalookup" without arguments (which will fill in the data for all your recordings), then going to the PBB and looking at that field.  Once you've assigned inetrefs to all your recording rules, the accuracy should go way up (though it should work very well without doing so).