Fix MythUI File Browser.
It appears that with recent changes to SG file listing methods, the returned files are often/usually prepended with a slash, and having a interstitial // seems to throw remotefile into a fit.  This just adjusts the MythUIfileBrowser to work again so that I can get on with using it.  I'll look for a more elegant solution once I finish my current tasks.

Fixes #9895.