Add shell class for maintaining outbound connections to the master server within the mechanics of libmythprotoserver.  This is intended to receive inbound commands and events, used to manage slave backends, mediaservers, jobqueues, and any future application the master backend needs direct control over.
In order to use, the virtual AnnounceSocket should be overridden and
send the proper announce string to the master backend using the supplied
socket. The ConnectToMaster() method is used to create the initial
connection, and must not be called until after the request handler is
registered to the socket manager. After which, the socket will
automatically attempt reconnection every five seconds if it gets
disconnected.