Remove offending putchar(".") from UPnP autoconf
These .s spew later on when the stdout gets flushed, which makes a mess.

Also, moved the delay to the top of the loop to allow more time for multiple
backends to reply.  If you have two backends and one is significantly faster
than the other replying, there was a pretty good chance that you'd only see
the fast one.