Improve check for whether to use OpenGL painter.
Move the code that decides whether to use the OpenGL painter to
MythRenderOpenGL, where it can check additional features.  Also, add a
check that causes fallback to the Qt painter when the OpenGL Renderer
contains the string "Software Rasterizer" even when the QGLFormat says
direct rendering is enabled, which was seen on a system using XFCE with
Compositor and X forwarding.

Any users who were previously overriding the OpenGL auto-detection
should try auto-detection again without overriding to see if we're now
properly detecting that OpenGL performance is not sufficient on their
systems.  And, users who still need to override the auto-detection after
this change can provide default-verbosity logs showing startup of
mythtv-setup or mythfrontend so we can attempt to detect other
indicators of bad hardware/driver combinations.

Thanks to hadack on IRC for providing the logs and testing required.