Rework ordering of command line processing in mythcommflag.  Inputs are no longer processed and stored to temporary variables, as required by the previous command line parser, but are now pulled out of the parser object as needed. The --skipdb parameter now behaves properly.
Fixes #9515

Use with no options is no longer permitted.  The old behavior of running
all recordings directly is no longer supported.  If the --queue
parameter is specified alone, it will queue a commercial detection job
for any recording that has not been previously flagged, and is not on a
commercial free channel.  The --force parameter will queue jobs for all
recordings, and the --rebuild will queue jobs to instead rebuild the
seektables.

Fixes #9152