Convert all text files from CRLF to LF (DOS->UNIX)
Seems that git doesn't enforce UNIX file format unless you specifically tell
it to.  Following the instructions at:
http://help.github.com/line-endings/
I have converted them all to UNIX file format (LF only)

It is strongly recommended for all developers to follow their configuration
advice on this matter to keep the text files properly configured.  This includes
setting to yes on Windows boxes so it will convert to CRLF on checkout and back
to LF on checkin.