Recording Metadata: Allow editing season, episode, and inetref from the PBB.
You can now edit the season, episode, and inetref of a recording in the Watch Recordings screen:

MENU->Recording Options->Change Recording Metadata

Really, though, this should ONLY be done when a run of "mythmetadatalookup" without arguments doesn't set the inetref, season, or episode properly.  This should be under 10% of the time, well under 10% of the time if you've set up the inetrefs in your recording rules.

Much as I am loath to do it, this adds new required fields to the editmetadata window in the PBB.  As ever, you can get the most up-to-date version of Arclight for the theme I've made these changes on (as opposed to adding them to other themes as an afterthought):

http://www.fecitfacta.com/Arclight.tar.gz