Change VERBOSE_LEVEL_CHECK to use both mask and level
Added a new pseudo-level of LOG_ANY.  This allows you to check for the mask
being set, and the level set to anything.  However, should you try to log a
message at that level, it will be summarily discarded.

This concludes the conversion of logging, unless we find mismatched levels,
which will be dealt with on a case by case basis.  I intend to still sweep
through the changes to verify the conversion, but if you find ones that seem
to be mismatched, either let me know, or just fix them yourself, whichever
works.