MythScreenType: Fix the area of a pop-up.
We use a position of (-1, -1) to indicate that a pop-up should be centered.
To center the pop-up we need to move the whole rectangle, not
just the edge.  Qt's setX and setY position the edge of the rectangle,
leaving the other edges alone.  To move the whole rectangle, we need to use
QRect's move functions, instead of the set functions.

This fixes sub-elements use of percentages, since the area is now correct.