MythMetadataLookup: Copy recording rule inetrefs to recordings.
If a mass update is run (--refresh-all, --refresh-all-rules, or no arguments), attempt to populate inetrefs bidirectionally.

With --refresh-all, see if a inetref'less program's recording rule has an inetref.  If it does, copy it across before performing a lookup.

With --refresh-all-rules, after looking up the rules themselves, copy newly-found inetrefs to the affected recordings.