Web Setup: Make the xmltv edit in the channel editor an autocomplete edit.
The xmltv edit will now present a list of xmltvid's that are valid for the
source of the channel you are editing (pulled in from the *.xmltv file for
the source). The list is filtered as you type into the edit.

Currently only useful for xmltv users but could probably be extended for
schedules direct if there is interest.