Recording Metadata: Allow picking artwork for individual seasons on recordings.
http://www.fecitfacta.com/recimages.ogv

If you really want to see a simple summary of what I'm about to say, just watch the video above.

In the Recording Rules Editor, under the Metadata Options section, you can now select your fanart, coverart, and banners for recording rules by season.  These values now persist and don't require an expensive hunt through the filesystem to look for.

Simply, each season + inetref (or in the case of a movie, just an inetref) has a row in the new recordedartwork table.  So, as you change the season values in the screen, the artwork being modified changes.  If you want to control what artwork appears for Season 4 of True Blood, edit the "True Blood" recording rule and set the season number to 4.  Now, start browsing local and online file sources and set the artwork you like for that season.  These changes take place immediately upon selection.

This is obviously a more interactive and time consuming process than the old "run a script that does its best in the background" method, but it's also 100% *your* choice what artwork you get, it supports different artwork for multiple seasons, etc.  Well worth it IMO.

If a given season doesn't have new artwork assigned to it yet, it will fall back to the latest seasons artwork.

I have *not* yet removed the image hunt from the Watch Recordings screen.  This is the last element that remains in the transition to the new recording metadata code.  Almost there!

As ever, version of Arclight to match is found here:

http://www.fecitfacta.com/Arclight.tar.gz