PlaybackBox: Remove Image Hunt in favor of user-selected artwork.
(and incur the ire of those who think the image hunt is "good enough")

http://www.mythtv.org/wiki/Enhancing_Recordings_with_Metadata_Lookup#Picking_Artwork_for_Your_Recording_Rules

As documented above, remove the costly, slow image hunt and instead allow the user to assign exactly the images they want, to exactly which recording rules, and even to particular seasons within that rule.

I will work on ways to pre-populate recording rules in the coming days.  Possibly a one-time set of image hunts in mythmetadatalookup that fills the recordedartwork table.