Changed the replex code to use mythlogging
This should work fine in the built-in case, although the output of the
mythreplex case may look a bit interesting...  as the old code had a pile of
cases where the easiest conversion was to multiple log lines.  If it looks too
hideous, I'll redo it to create a static buffer then log from it.

Refs #9903

If you see issues or missed output that should go to the logs, please let me
know.  If I don't hear back in a week or so, I'll likely close the ticket.