Fix log messages containing %
Fixed #9919

The issue here is that after the fully formatted QString is converted to a
char *, it is still run through vsnprintf as that is common code with the C
implementation.  To fix this, I have added a regexp-based replace in the macro
that will convert "%" or "%%" to "%%".  This will account for places we already
put "%%" into the QString, and for those we did not.