Housekeeper: Perform a nightly update of artwork for recordings and rules.
This is somewhat analogous to the JAMU -MW cronjob many users have used for a couple of years, but more powerful.  Once every 24 hours, the backend will check to see if any new artwork is needed and download it... for example, if a new season of a show starts.  The next day, you'll have all your old artwork, assigned to the old seasons, and the artwork for the new season as well, assigned to episodes thereafter.

This will be most effective if you've set inetrefs on your recordings or at *least* run "mythmetadatalookup --refresh-all-rules; mythmetdatalookup --refresh-all" once, since it only operates on rules/recordings with inetrefs.