Further tweak recording metadata logic.
This refined logic prevents a image update for a movie inetref from first trying it in the television grabber.

Previously, you could set an inetref on a movie, and it would STILL iterate through the TV grabber when trying to grab artwork, then only after a failure try the Movie grabber.  That logic could be better.  Instead, if we are handling images, we are at the very end of the metadata chain, and thus we need to trust the data we've got, one way or another. If there's an inetref, and season and episode are > 0, this is a TV show.  If there's an inetref and nothing else, this is a movie.