Add an extra fallback lookup for TV Shows.
When a rec rule has title and subtitle, this was the lookup that was used.  However, this relies on the title to be an exact match.  If there ARE a title and subtitle, try one last time by dropping the subtitle.

Makes a lookup for "Castle" return a selection list when the program being looked up has a subtitle.