Prevent unwanted lookup of generic episodes.
Check the recording we're about to grab artwork for.  If it has an inetref, but no season, episode, or subtitle, yet its recording rule does have a season/episode, it's a generic TV show.  We can't possibly get artwork of value for this, so don't try (in non-dangerous mode, anyway-- dangerous mode will still try, but probably fail at it).