ProgramInfo: Add a Query method to get Category_type.
Not generally useful for all applications of programinfo, thus I didn't add it.  Helpful to me for metadata lookup, though, as I can use the category type as a hint for which grabber to start with.  It's definitely not a trustworthy source of information, but it can help break a tie.

Bump library API.