Backend Video Scanner: Perform all metadata updates and image download the UI can.
Replaces the old video scanner thread in the backend (well, old is relative) and replace it with a MetadataFactory.

Add a video scanning thread to the Metadata Factory, and handling for video metadata in an automated lookup scenario.

The hardest part and largest bits of this change are actually code to make the metadata download code agnostic to the fact that it's running on a backend.  Specifically, it all makes liberal use of RemoteFile, which will function, but spewed nasty errors about the backend connecting to itself.  So, lots and lots of code to detect and work around that.

So, to all who have ever complained that they couldn't script a scan, or used it as an excuse for using some script that molested and destroyed their data, please, please instead use:

mythbackend --scanvideos

or the SCAN_VIDEOS protocol command.  This now does everything you can from the frontend, and uses the same classes, so it will never need updating when the frontend gains some new scanning functionality.