Convert MythHostAddressAny() to return QHostAddress
As everywhere it is used, we immediately convert it to QHostAddress, it only
makes sense to make the method return it already converted.  Also, using the
SpecialAddress case in the initialization of the value makes it more Qt-ish.

Changing this required adding QT = network into .pro files for mythjobqueue
and mytharchivehelper so it could find the QHostAddress include under
qt4/QtNetwork.