Put channel changing script handling back to MythSystem
Fixes #9905

It seems that the merge of mythtv-rec2 branch that happened in May effectively
reverted the change of this section of code to use the new MythSystem support.
See 9b22460f53bf6313516f33987e4260897c6076a3

This reimplements the use of MythSystem rather than QProcess while leaving as
much of the rest alone as possible.  It also removes a now unneeded thread.

Also cleaned up a stray include of QProcess in the grabber settings.