Make dummy recordings send REC_FINISHED event
A dummy recording (such as during livetv while tuning the device) sends a
REC_STARTED event, but not a REC_FINISHED event.  This was causing issues with
my event scripts which I was trying to make abort if the recording finished
while I was waiting.  Having the behavior matched makes it more useful.