Metadata: Add XML export methods.
Adds some methods to transform a ProgramInfo into a MetadataLookup, which is the common denominator between Recordings, Video, Music, and Games, and can be exported or reimported.

Once the UI and/or scanner is there, you could technically export a Recording, then drop it in MythVideo w/ the XML file and have all the metadata brought back in, or vice versa.