Metadata: Prevent 1x01-on-every-recording with title-only matches.
We use a special handling type when something is a TV show but we only have generic results.  In the UI, we want to handle it one way, but in mythmetadatalookup, we want to not set season and episode when we've utilized this generic/title-only handling type.

Should clear up some inconsistencies (I introduced this bug just a few days ago, so hopefully not too many).