Metadata: Attempt to match based on year when matching returns > 1 result.
In mythmetadatalookup and the schedule editor, when multiple results are returned for, say, "Clash of the Titans," check to see if any of the results have the same year as the OriginalAirdate Year or Airdate year.  If so, pick that one rather than presenting the list.  My very limited testing of this was fairly flawless, although obviously the possibility for error exists (Original V. Special edition or director's cut).  If there are *multiple* results within a single year, fall back to the regular method of presenting a multiresult dialog box.

Also fix a number of other small bugs (like movies just grabbing the first of a multiresult list).