Code cleanup in GuideGrid
Remove ~80 lines of code duplication by refactoring.

Add templatable starttime/endtime strings to the time list so that
constructs such as "20:30-21:00" are possible.

Remove overloading of the datetext textarea as this cannot work well
with templating or layouts which assign a precise amount of space. If
necessary we can add it back without overloading a field which outwardly
purports to show a date.