Make "play program" network control call synchronous.
This moves the response from "play program" while inside the playbackbox
into the handling thread, rather than a separate asynchronous result.
This may result in some concurrency issues as the entire network control
server is locked until that thread to that client responds. Due to the
already very coarse locking, it should not be significantly noticeable.

Fixes #9922