Common date/time to string conversion functions
We perform these conversions all over the place which makes
maintenance and new date/time features harder to implement. Existing
code will need to be converted to use these instead and this commit
doesn't address that.

Format can be indicated via the kDateFull, kDateShort, kDateTimeFull,
kDateTimeShort or kTime flags.

The first new feature is the option to 'simplify' dates to 'Today',
'Tomorrow' or 'Yesterday' when applicable. This can be specified with
the kSimplify flag