Added support for max queue length for db logging
It the DB should not be available, we requeue the logging messages that were
to be sent to it.  Rather than queue without bound should the DB take a
vacation, we will now only queue up to 1000 messages, and then will disable
queuing until such a time that the DB is once again available.

Also, rearranged when log items are deleted from memory, possibly thwarting
a memory leak.